Allocations: link to BudgetDetail
This required adding `createHref` to AragonHistory

I also moved AragonHistory to its own file, since api-react/index was
becoming unwieldy.
