+++ apps/address-book/package.json
@@ -40,6 +40,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "styled-components": "4.1.3",
     "web3-utils": "^1.0.0"
   },
