+++ apps/dot-voting/package.json
@@ -43,6 +43,7 @@
     "react-dom": "^16.8.6",
     "react-motion": "0.5.2",
     "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "styled-components": "4.1.3",
     "web3-utils": "^1.0.0"
   },
